[SASS-5595: Renamed controllers, views and tests + refactored logic to show messages]
