+++ test/views/offPayrollWorking/EmployerTaxWarningViewSpec.scala
@@ -22,9 +22,9 @@ import org.jsoup.nodes.Document
 import play.api.i18n.Messages
 import play.api.mvc.AnyContent
 import support.ViewUnitTest
-import views.html.offPayrollWorking.UkTaxTakenFromPayView
+import views.html.offPayrollWorking.EmployerTaxWarningView
 
-class UkTaxTakenFromPayViewSpec extends ViewUnitTest {
+class EmployerTaxWarningViewSpec extends ViewUnitTest {
   object Selectors {
     val paragraph1Selector = "#main-content > div > div > p:nth-child(2)"
     val bullet1Selector: String = "#main-content > div > div > ul > li"
@@ -91,9 +91,9 @@ class UkTaxTakenFromPayViewSpec extends ViewUnitTest {
     UserScenario(isWelsh = true, isAgent = true, CommonExpectedCY, Some(ExpectedAgentCY))
   )
 
-  private lazy val underTest = inject[UkTaxTakenFromPayView]
+  private lazy val underTest = inject[EmployerTaxWarningView]
 
-  private val cancelUrl = s"http://localhost:9302/update-and-submit-income-tax-return/$taxYear/income-tax-return-overview" //TODO: This needs to be changed when Samuel's ticket is merged
+  private val cancelUrl = s"http://localhost:9302/update-and-submit-income-tax-return/$taxYear/income-tax-return-overview"
   private val continueUrl = s"http://localhost:9302/update-and-submit-income-tax-return/$taxYear/income-tax-return-overview"
 
   userScenarios.foreach { userScenario =>
@@ -104,13 +104,7 @@ class UkTaxTakenFromPayViewSpec extends ViewUnitTest {
         implicit val authRequest: AuthorisationRequest[AnyContent] = getAuthRequest(userScenario.isAgent)
         implicit val messages: Messages = getMessages(userScenario.isWelsh)
 
-        val htmlFormat = underTest(taxYear,
-          s"employment.employerOpw.warning.tax.title.${if (userScenario.isAgent) "agent" else "individual"}",
-          s"employment.employerOpw.warning.tax.heading.${if (userScenario.isAgent) "agent" else "individual"}",
-          s"employment.employerOpw.warning.tax.p2.${if (userScenario.isAgent) "agent" else "individual"}",
-          continueUrl,
-          cancelUrl
-        )
+        val htmlFormat = underTest(taxYear, continueUrl, cancelUrl)
 
         implicit val document: Document = Jsoup.parse(htmlFormat.body)
